Compose DB out of data stored externally
