+++ .travis.yml
@@ -13,6 +13,7 @@ matrix:
 before_install:
   - bash scripts/setup-osx.sh
 install:
+  - python3 db_update.py
   - export PYFA_VERSION="$(python3 scripts/dump_version.py)"
   - bash scripts/package-osx.sh
 before_deploy:
