+++ pyfa.py
@@ -26,6 +26,7 @@ from optparse import AmbiguousOptionError, BadOptionError, OptionParser
 
 import config
 from service.prereqsCheck import PreCheckException, PreCheckMessage, version_block, version_precheck
+from db_update import db_needs_update, update_db
 
 
 # ascii_text = '''
@@ -115,6 +116,10 @@ if __name__ == "__main__":
             pyfalog.info("Running in a frozen state.")
         else:
             pyfalog.info("Running in a thawed state.")
+
+        if db_needs_update() is True:
+            pyfalog.info("Gamedata needs an update")
+            update_db()
 
         # Lets get to the good stuff, shall we?
         import eos.db
