+++ staticdata/fsd_binary/metagroups.json
@@ -1,0 +1,112 @@
+{
+  "1": {
+    "name": "Tech I", 
+    "nameID": 66672
+  }, 
+  "2": {
+    "iconID": 24150, 
+    "iconSuffix": "tech2", 
+    "name": "Tech II", 
+    "nameID": 66673
+  }, 
+  "3": {
+    "iconID": 24147, 
+    "iconSuffix": "storyline", 
+    "name": "Storyline", 
+    "nameID": 66674
+  }, 
+  "4": {
+    "iconID": 24146, 
+    "iconSuffix": "faction", 
+    "name": "Faction", 
+    "nameID": 66675
+  }, 
+  "5": {
+    "iconID": 24149, 
+    "iconSuffix": "officer", 
+    "name": "Officer", 
+    "nameID": 66676
+  }, 
+  "6": {
+    "description": "Modules found in deadspace.", 
+    "descriptionID": 66671, 
+    "iconID": 24148, 
+    "iconSuffix": "deadspace", 
+    "name": "Deadspace", 
+    "nameID": 66677
+  }, 
+  "7": {
+    "name": "Frigates", 
+    "nameID": 66678
+  }, 
+  "8": {
+    "name": "Elite Frigates", 
+    "nameID": 66679
+  }, 
+  "9": {
+    "name": "Commander Frigates", 
+    "nameID": 66680
+  }, 
+  "10": {
+    "name": "Destroyer", 
+    "nameID": 66681
+  }, 
+  "11": {
+    "name": "Cruiser", 
+    "nameID": 66682
+  }, 
+  "12": {
+    "name": "Elite Cruiser", 
+    "nameID": 66683
+  }, 
+  "13": {
+    "name": "Commander Cruiser", 
+    "nameID": 66684
+  }, 
+  "14": {
+    "iconID": 24151, 
+    "iconSuffix": "tech3", 
+    "name": "Tech III", 
+    "nameID": 66685
+  }, 
+  "15": {
+    "iconID": 24152, 
+    "iconSuffix": "abyssal", 
+    "name": "Abyssal", 
+    "nameID": 317771
+  }, 
+  "17": {
+    "description": "This item is only available through the New Eden Store or exclusive offers.", 
+    "descriptionID": 317894, 
+    "iconID": 24153, 
+    "iconSuffix": "premium", 
+    "name": "Premium", 
+    "nameID": 317893
+  }, 
+  "19": {
+    "description": "This item is only available for a limited time.", 
+    "descriptionID": 317896, 
+    "iconID": 24154, 
+    "iconSuffix": "limited", 
+    "name": "Limited Time", 
+    "nameID": 317895
+  }, 
+  "52": {
+    "iconID": 24155, 
+    "iconSuffix": "structureFaction", 
+    "name": "Structure Faction", 
+    "nameID": 550638
+  }, 
+  "53": {
+    "iconID": 24156, 
+    "iconSuffix": "structureTech2", 
+    "name": "Structure Tech II", 
+    "nameID": 550639
+  }, 
+  "54": {
+    "iconID": 24157, 
+    "iconSuffix": "structureTech1", 
+    "name": "Structure Tech I", 
+    "nameID": 550644
+  }
+}
