+++ staticdata/fsd_lite/evecategories.json
@@ -1,0 +1,273 @@
+{
+  "0": {
+    "categoryID": 0, 
+    "categoryName": "#System", 
+    "categoryNameID": 63539, 
+    "published": false
+  }, 
+  "1": {
+    "categoryID": 1, 
+    "categoryName": "Owner", 
+    "categoryNameID": 63540, 
+    "published": false
+  }, 
+  "10": {
+    "categoryID": 10, 
+    "categoryName": "Trading", 
+    "categoryNameID": 63548, 
+    "published": false
+  }, 
+  "11": {
+    "categoryID": 11, 
+    "categoryName": "Entity", 
+    "categoryNameID": 63549, 
+    "published": false
+  }, 
+  "14": {
+    "categoryID": 14, 
+    "categoryName": "Bonus", 
+    "categoryNameID": 63550, 
+    "iconID": 0, 
+    "published": false
+  }, 
+  "16": {
+    "categoryID": 16, 
+    "categoryName": "Skill", 
+    "categoryNameID": 63551, 
+    "iconID": 33, 
+    "published": true
+  }, 
+  "17": {
+    "categoryID": 17, 
+    "categoryName": "Commodity", 
+    "categoryNameID": 63552, 
+    "iconID": 0, 
+    "published": true
+  }, 
+  "18": {
+    "categoryID": 18, 
+    "categoryName": "Drone", 
+    "categoryNameID": 63553, 
+    "iconID": 0, 
+    "published": true, 
+    "sofBuildClass": "ship"
+  }, 
+  "2": {
+    "categoryID": 2, 
+    "categoryName": "Celestial", 
+    "categoryNameID": 63541, 
+    "published": true
+  }, 
+  "20": {
+    "categoryID": 20, 
+    "categoryName": "Implant", 
+    "categoryNameID": 63554, 
+    "iconID": 0, 
+    "published": true
+  }, 
+  "22": {
+    "categoryID": 22, 
+    "categoryName": "Deployable", 
+    "categoryNameID": 63555, 
+    "iconID": 0, 
+    "published": true
+  }, 
+  "23": {
+    "categoryID": 23, 
+    "categoryName": "Starbase", 
+    "categoryNameID": 63556, 
+    "iconID": 0, 
+    "published": true
+  }, 
+  "24": {
+    "categoryID": 24, 
+    "categoryName": "Reaction", 
+    "categoryNameID": 63557, 
+    "iconID": 0, 
+    "published": true
+  }, 
+  "25": {
+    "categoryID": 25, 
+    "categoryName": "Asteroid", 
+    "categoryNameID": 63558, 
+    "published": true
+  }, 
+  "26": {
+    "categoryID": 26, 
+    "categoryName": "WorldSpace", 
+    "categoryNameID": 63568, 
+    "published": false
+  }, 
+  "29": {
+    "categoryID": 29, 
+    "categoryName": "Abstract", 
+    "categoryNameID": 63559, 
+    "published": false
+  }, 
+  "3": {
+    "categoryID": 3, 
+    "categoryName": "Station", 
+    "categoryNameID": 63542, 
+    "published": false
+  }, 
+  "30": {
+    "categoryID": 30, 
+    "categoryName": "Apparel", 
+    "categoryNameID": 63572, 
+    "published": true
+  }, 
+  "32": {
+    "categoryID": 32, 
+    "categoryName": "Subsystem", 
+    "categoryNameID": 63562, 
+    "published": true
+  }, 
+  "34": {
+    "categoryID": 34, 
+    "categoryName": "Ancient Relics", 
+    "categoryNameID": 63561, 
+    "published": true
+  }, 
+  "35": {
+    "categoryID": 35, 
+    "categoryName": "Decryptors", 
+    "categoryNameID": 63563, 
+    "published": true
+  }, 
+  "350001": {
+    "categoryID": 350001, 
+    "categoryName": "Infantry", 
+    "categoryNameID": 267649, 
+    "published": false
+  }, 
+  "39": {
+    "categoryID": 39, 
+    "categoryName": "Infrastructure Upgrades", 
+    "categoryNameID": 63565, 
+    "published": true
+  }, 
+  "4": {
+    "categoryID": 4, 
+    "categoryName": "Material", 
+    "categoryNameID": 63543, 
+    "iconID": 22, 
+    "published": true
+  }, 
+  "40": {
+    "categoryID": 40, 
+    "categoryName": "Sovereignty Structures", 
+    "categoryNameID": 63564, 
+    "published": true
+  }, 
+  "41": {
+    "categoryID": 41, 
+    "categoryName": "Planetary Interaction", 
+    "categoryNameID": 63569, 
+    "published": true
+  }, 
+  "42": {
+    "categoryID": 42, 
+    "categoryName": "Planetary Resources", 
+    "categoryNameID": 63566, 
+    "published": true
+  }, 
+  "43": {
+    "categoryID": 43, 
+    "categoryName": "Planetary Commodities", 
+    "categoryNameID": 63567, 
+    "published": true
+  }, 
+  "46": {
+    "categoryID": 46, 
+    "categoryName": "Orbitals", 
+    "categoryNameID": 63570, 
+    "published": true
+  }, 
+  "49": {
+    "categoryID": 49, 
+    "categoryName": "Placeables", 
+    "categoryNameID": 63571, 
+    "published": false
+  }, 
+  "5": {
+    "categoryID": 5, 
+    "categoryName": "Accessories", 
+    "categoryNameID": 63560, 
+    "iconID": 33, 
+    "published": true
+  }, 
+  "53": {
+    "categoryID": 53, 
+    "categoryName": "Effects", 
+    "categoryNameID": 63573, 
+    "published": false
+  }, 
+  "54": {
+    "categoryID": 54, 
+    "categoryName": "Lights", 
+    "categoryNameID": 63574, 
+    "published": false
+  }, 
+  "59": {
+    "categoryID": 59, 
+    "categoryName": "Cells", 
+    "categoryNameID": 235965, 
+    "published": false
+  }, 
+  "6": {
+    "categoryID": 6, 
+    "categoryName": "Ship", 
+    "categoryNameID": 63544, 
+    "published": true
+  }, 
+  "63": {
+    "categoryID": 63, 
+    "categoryName": "Special Edition Assets", 
+    "categoryNameID": 285070, 
+    "published": true
+  }, 
+  "65": {
+    "categoryID": 65, 
+    "categoryName": "Structure", 
+    "categoryNameID": 308338, 
+    "published": true
+  }, 
+  "66": {
+    "categoryID": 66, 
+    "categoryName": "Structure Module", 
+    "categoryNameID": 308340, 
+    "published": true
+  }, 
+  "7": {
+    "categoryID": 7, 
+    "categoryName": "Module", 
+    "categoryNameID": 63545, 
+    "iconID": 67, 
+    "published": true
+  }, 
+  "8": {
+    "categoryID": 8, 
+    "categoryName": "Charge", 
+    "categoryNameID": 63546, 
+    "published": true
+  }, 
+  "87": {
+    "categoryID": 87, 
+    "categoryName": "Fighter", 
+    "categoryNameID": 510368, 
+    "published": true
+  }, 
+  "9": {
+    "categoryID": 9, 
+    "categoryName": "Blueprint", 
+    "categoryNameID": 63547, 
+    "iconID": 21, 
+    "published": true
+  }, 
+  "91": {
+    "categoryID": 91, 
+    "categoryName": "SKINs", 
+    "categoryNameID": 531338, 
+    "published": true
+  }
+}
