+++ staticdata/phobos/metadata.json
@@ -1,0 +1,10 @@
+[
+  {
+    "field_name": "client_build", 
+    "field_value": 1595463
+  }, 
+  {
+    "field_name": "dump_time", 
+    "field_value": 1572435719
+  }
+]
